Handle missing font
For situations where `fontconfig` is installed but no suitable font is
found.
One such case might be when compiling inside a chroot where no fonts
are present.

Without some kind of fallback, the program will happily compile with an
empty UI_FONT_PATH.
This leads to a segfault when trying to run `gf2`.