Use %-formatting in __gf_fields_recurse
My version of GDB is so old, it uses Python 2. In Python 2
`print('(array)', 5)` results in `('(array)', 5)` being printed
instead of `(array) 5` which breaks the parsing in `WatchAddFields`.

Since %-formatting works the same in all of the versions of Python it
is relatively easy to add and fix the problem for older setups.